Extract the tests for System.MemInfo into MemInfoSpec
